Refactor(#51): dto 폴더 위치 오류 수정
