+++ src/main/java/com/numberone/backend/domain/member/contoller/MemberController.java
@@ -1,7 +1,7 @@
 package com.numberone.backend.domain.member.contoller;
 
-import com.numberone.backend.domain.member.dto.dto.request.BuyHeartRequest;
-import com.numberone.backend.domain.member.dto.dto.response.HeartCntResponse;
+import com.numberone.backend.domain.member.dto.request.BuyHeartRequest;
+import com.numberone.backend.domain.member.dto.response.HeartCntResponse;
 import com.numberone.backend.domain.member.service.MemberService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
